+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Provided functionality to disable interactivity using an environment variable `AZUREAUTH_DISABLE_INTERACTIVITY`.
+
 ## [0.4.0] - 2022-06-23
 ### Added
 - Environment variable `AZUREAUTH_CACHE` and option `--cache` to support a custom cache location on Windows.